+++ scripts/macros/generic.mjs
@@ -10,3 +10,4 @@ export {advancedMeleeAttack} from './generic/advancedMeleeAttack.mjs';
 export {movementAnimation} from './generic/movementAnimation.mjs';
 export {checkModifier, saveModifier, skillModifier, toolModifier} from './generic/rollModifier.mjs';
 export {additionalTargets} from './generic/additionalTargets.mjs';
+export {damageBonusToOneRoll} from './generic/damageBonusToOneRoll.mjs';