+++ scripts/macros/generic/damageBonusToOneRoll.mjs
@@ -1,0 +1,136 @@
+import {automationUtils, constants, DamageBonus, documentUtils, workflowUtils} from '../../proxy.mjs';
+async function damage({document, workflow}) {
+    if (!workflow.targets.size) return;
+    const config = automationUtils.getGenericConfigValues(document, 'chris-premades', 'damageBonusToOneRoll', Object.keys(damageBonusToOneRoll.genericConfig));
+    if (!config.bonus.length) return;
+    if (config.attackType.length) {   
+        if (!workflowUtils.isAttackType(workflow, config.attackType)) return;
+    }
+    if (config.damageType.length) {
+        const damage = workflow.activity.damage ?? workflow.activity.otherActivity?.damage;
+        if (!damage) return;
+        if (!damage.parts.some(p => config.damageType.some(d => p.types.has(d)))) return;
+    }
+    if (config.healingType.length) {
+        const heal = workflow.activity.healing;
+        if (!heal) return;
+        if (!config.healingType.some(d => heal.types.has(d))) return;
+    }
+    if (config.identifiers.length) {
+        if (!config.identifiers.includes(documentUtils.getIdentifier(workflow.item))) return;
+    }
+    if (config.itemType.length) {
+        if (!config.itemType.includes(workflow.item.type)) return;
+    }
+    if (config.spellSchool.length) {
+        if (!config.spellSchool.includes(workflow.item.system.school)) return;
+    }
+    const multiSingleTarget = workflow.workflowOptions['chris-premades']?.multiSingleTarget;
+    if (multiSingleTarget && document.flags['chris-premades']?.lastUse === multiSingleTarget.rollID) return;
+    const optional = multiSingleTarget ? multiSingleTarget.remainingAttacks > 1 : false;
+    return new DamageBonus(document, {formula: config.bonus, optional})
+        .withOnUse(async ({bonus}) => {
+            if (multiSingleTarget) await documentUtils.setFlag(document, 'chris-premades', 'lastUse', multiSingleTarget.rollID);
+            const item = bonus.document.documentName === 'Item' ? bonus.document : bonus.activity?.item;
+            if (!item) return;
+            if (config.rollItem) {
+                await workflowUtils.completeItemUse(item, Array.from(bonus.targets ?? []));
+            } else if (config.rollActivity) {
+                const activity = item.system.activities.get(config.rollActivity);
+                if (!activity) return;
+                await workflowUtils.completeActivityUse(activity, Array.from(bonus.targets ?? []));
+            }
+        })
+        .initialize();
+}
+export const damageBonusToOneRoll = {
+    rules: 'all',
+    version: '2.0.3',
+    category: 'damage',
+    generic: true,
+    documents: ['activeeffect', 'item'],
+    roll: [
+        {
+            pass: 'actorOptionalBonusDamage',
+            macro: damage,
+            priority: 250
+        }
+    ],
+    genericConfig: {
+        attackType: {
+            default: '',
+            type: 'select',
+            category: 'behavior',
+            label: 'CHRISPREMADES.Config.AttackType.Label',
+            hint: 'CHRISPREMADES.Macros.Generic.Common.AttackTypeHint',
+            get options() { return [
+                'attack',
+                'meleeAttack',
+                'rangedAttack',
+                'weaponAttack',
+                'spellAttack',
+                'rangedWeaponAttack',
+                'meleeWeaponAttack',
+                'rangedSpellAttack',
+                'meleeSpellAttack'
+            ].map(a => ({value: a, label: _loc('CHRISPREMADES.Config.AttackType.' + a)})); }
+        },
+        bonus: {
+            default: '',
+            type: 'text',
+            label: 'CHRISPREMADES.Config.DamageBonus',
+            category: 'behavior'
+        },
+        damageType: {
+            default: [],
+            type: 'select-many',
+            category: 'behavior',
+            label: 'CHRISPREMADES.Config.DamageType',
+            hint: 'CHRISPREMADES.Macros.Generic.Common.DamageTypeHint',
+            get options() { return constants.damageTypeOptions(); }
+        },
+        healingType: {
+            default: [],
+            type: 'select-many',
+            category: 'behavior',
+            label: 'CHRISPREMADES.Config.HealingType',
+            hint: 'CHRISPREMADES.Macros.Generic.Common.HealingTypeHint',
+            get options() { return constants.healingTypeOptions(); }
+        },
+        identifiers: {
+            default: [],
+            type: 'selectIdentifiers',
+            category: 'behavior',
+            label: 'CHRISPREMADES.Config.Identifiers',
+            hint: 'CHRISPREMADES.Macros.Generic.Common.IdentifierHint'
+        },
+        itemType: {
+            default: [],
+            type: 'select-many',
+            category: 'behavior',
+            label: 'CHRISPREMADES.Config.ItemTypes',
+            hint: 'CHRISPREMADES.Macros.Generic.Common.ItemTypeHint',
+            get options() { return constants.usableItemTypes(); }
+        },
+        rollActivity: {
+            default: '',
+            type: 'selectActivity',
+            category: 'behavior',
+            label: 'CHRISPREMADES.Macros.Generic.Common.RollActivity'
+        },
+        rollItem: {
+            default: false,
+            type: 'checkbox',
+            category: 'behavior',
+            label:'CHRISPREMADES.Macros.Generic.Common.RollItem'
+        },
+        spellSchool: {
+            default: [],
+            type: 'select-many',
+            category: 'behavior',
+            label: 'CHRISPREMADES.Config.SpellSchool',
+            hint: 'CHRISPREMADES.Macros.Generic.Common.SpellSchoolHint',
+            get options() { return constants.spellSchoolOptions(); }
+        }
+    }
+};